refactor: config 내부의 validate 메서드를 별도 클래스로 분리 (#1)
