+++ hwpxtemplater-core/src/main/java/kr/mumberrymountain/hwpxtemplater/validate/Validater.java
@@ -1,0 +1,7 @@
+package kr.mumberrymountain.hwpxtemplater.validate;
+
+import java.util.Map;
+
+public interface Validater {
+    public void validate(Map<String, Object> properties);
+}
