+++ hwpxtemplater-core/src/main/java/kr/mumberrymountain/hwpxtemplater/validate/ValidaterFactory.java
@@ -1,0 +1,13 @@
+package kr.mumberrymountain.hwpxtemplater.validate;
+
+import java.util.Arrays;
+import java.util.List;
+
+public class ValidaterFactory {
+    public static List<Validater> validaters() {
+        return Arrays.asList(
+                DelimValidater.getInstance(),
+                CharRoleSetterValidater.getInstance()
+        );
+    }
+}
